Update code.js
Realized I could edit the search function already in our javascript file to make every text result in colorList hyperlink to the contacts page. It already iterated over the converted or "parsed" json object (now a javascript object) from the server and adds the text to the html element colorList. 

Before this I declared a variable that created an  "a" element with an href attribute containing the link to contacts. It's innerHTML (The text between the element's opening and closing tags) contained formatted text that prefaced each desired value we wanted to display from the javascript object. I then added this finalized element to the colorList element through its outerHTML (which is a reference to the entire HTML element created, from its opening and closing tags).